backend/requirements.txt
Flask
flask-cors
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
gunicorn
python-dotenv
